Refactored the model training and testing files
